Fixed potential error in get_masked_data

+++ qap/qap_utils.py
@@ -71,7 +71,7 @@ def read_json(json_filename):
     if not os.path.exists(json_filename):
         err = "\n\n[!] The JSON file provided does not exist.\nFilepath: " \
               "%s\n\n" % json_filename
-        raise_smart_exception(locals(),err)
+        raise_smart_exception(locals(), err)
 
     with open(json_filename, "r") as f:
         json_dict = json.load(f)
@@ -240,8 +240,8 @@ def get_masked_data(data, mask):
         mask_img = nb.load(mask)
         mask = mask_img.get_data()
 
-    data = np.asarray(data)
-    mask = np.asarray(mask)
+    data = np.asarray(data, dtype=float)
+    mask = np.asarray(mask, dtype=float)
 
     if len(data.shape) > 3:
         # if timeseries
